Move validation_denied_kind_to_string from service to graphql
