+++ graphql/src/schema/queries/me.rs
@@ -1,10 +1,9 @@
 use async_graphql::*;
-use service::permission_validation::{
-    has_api_role, validate, validation_denied_kind_to_string, ValidationError,
-};
+use service::permission_validation::{has_api_role, validate, ValidationError};
 use service::user_account::{UserAccount, UserAccountService};
 
 use crate::schema::types::{AccessDenied, DatabaseError, InternalError};
+use crate::schema::validation_denied_kind_to_string;
 use crate::ContextExt;
 
 use super::ErrorWrapper;
